Multiplayer on LAN
A few Fixes needed for multiplayer.
Mattresses and pillows need to become network objects. (They are only there for the player who spawned them.)
Toolbar canvas is not visible to client players.
Add a 3rd item that only when held allows for shooting.
Optional: Rotation of placeables? Holographic preview of placable?